인포러스트/criminal_project/requirements.txt
pip==9.0.3
Django==1.8.6
djangorestframework==3.3.1
gunicorn==19.1.1
jinja2==2.8
WhiteNoise==2.0.4
python-keystoneclient==2.0.0
MySQL-python==1.2.3
